rename: RecentLogService 위치 변경
