+++ src/main/java/com/example/tnote/boundedContext/home/controller/HomeController.java
@@ -8,13 +8,12 @@ import com.example.tnote.boundedContext.home.dto.ArchiveResponseDto;
 import com.example.tnote.boundedContext.home.dto.ArchiveSliceResponseDto;
 import com.example.tnote.boundedContext.home.dto.RecentLogResponseDto;
 import com.example.tnote.boundedContext.home.service.HomeService;
-import com.example.tnote.boundedContext.home.service.RecentLogService;
+import com.example.tnote.boundedContext.recentLog.service.RecentLogService;
 import com.example.tnote.boundedContext.observation.dto.ObservationResponseDto;
 import com.example.tnote.boundedContext.proceeding.dto.ProceedingResponseDto;
 import com.example.tnote.boundedContext.schedule.dto.SemesterNameResponseDto;
 import com.example.tnote.boundedContext.schedule.service.ScheduleService;
 import com.example.tnote.boundedContext.user.entity.auth.PrincipalDetails;
-import java.security.UnrecoverableEntryException;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
